+++ config.json
@@ -1,6 +1,7 @@
 {
-  "width": 800,
-  "height": 480,
-  "ps5_ip": "10.22.33.37",
-  "fullscreen": false
-}
+    "width": 800,
+    "height": 480,
+    "playstation_ip": "10.22.33.37",
+    "fullscreen": false
+  }
+
